requirements_pi.txt
gTTS
pyyaml
ping3
RPi.GPIO
